refactor: 전체 선택 컴포넌트 분리
